Save items in store, see explanation
This is a temporary solution! It will be changed as soon as possible.
It is bad design because how the item is mutated
is defined outside the store.
The main motivation was to migrate to API Version 2 with smallest
steps possible (keep the commits as short as possible).
Also that I do not get lost with all the changes made.

After the migration this solution will be removed.